fix(src/ex8): corrected np.mean to np.average
The fix consists in correcting the right numpy estatistical function to
calculate average bps.

Before:
The average bps was calculated using the arithmetic mean where every
symbol count as much as every other symbol. But that is not correct.

The way to calculate the average bps is through np.average (weighted
average) where different symbol count more to the average if their
probability is higher